Fix unrenamed bits in system units shipped in native packages
This is a bit of a SNAFU, because the service units are broken on fresh
systems, and on systems where EdgeDB 5.x and earlier was installed they
would work, but will initialize the data in the wrong spot
(`/var/lib/edgedb`) and owned by the wrong user (`edgedb`).  This fixes
the units and adds a check that errors out if the wrong data directory
is detected, so people who have successfully installed Gel 6.0/6.1 would
be informed to move and `chown` the data directory when upgrading to 6.2
and later.