Update setuptools to 77 for GelNoPostgres
It was originally pinned by #101.

I'm keeping a pin on a major verson, because upgrading major
version automatically might cause breakage in the future.

I need this update, because calver has published a broken
version, followed by a fix that bumps required setuptools.
Which means we cannot use the fix unless we update setuptools.