Improve performance of generic repo
The APT and RPM repos work on a locally-synced copy of their S3 prefixes
(because the repo tooling expects repos to be in a local fs), but the
generic repo currently just reads/writes things to S3 directly.  This is
problematic for atomicity and performance reasons.

Fix this by doing the S3 sync for the generic repo as well.  While at
it, lower log levels on some things to reduce noise.