Fix debian buster build
It is out of support, so point it at archive.debian.org instead.
We'll want to drop it, but probably not for 6.x.

Apparently we have support for casing on the variant in our template
files... I learned a lot more about sed today trying to understand
what update.sh was doing.